moved canvas components in a sub directory and found some more unused variables to remove
