+++ src/components/canvasParts/CSS/Canvas.css
@@ -15,7 +15,7 @@ body {
 }
 
 #canvas {
-    background-image: url(../../images/backgrounds/homseBaseSimplified2.png);
+    background-image: url(../../../images/backgrounds/homseBaseSimplified2.png);
     background-size: cover;
     background-repeat: no-repeat;
     background-attachment: fixed;
