+++ src/components/canvasParts/CSS/FireButton.css
@@ -1,6 +1,6 @@
 @font-face {
     font-family: call-of-ops-duty;
-    src: url(../../fonts/call-of-ops-duty-font/CallOfOpsDuty-qAyq.ttf);
+    src: url(../../../fonts/call-of-ops-duty-font/CallOfOpsDuty-qAyq.ttf);
 }
 
 :root {
